requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
dlib-bin==19.24.6
facenet-pytorch==2.5.3
filelock==3.19.1
fsspec==2025.7.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
requests==2.32.5
setuptools==80.9.0
sympy==1.14.0
torch==2.8.0
torchvision==0.23.0
typing_extensions==4.15.0
urllib3==2.5.0
